refactor: move flow logic into a flow directory
