+++ packages/backend/src/app/database/database-connection.ts
@@ -3,7 +3,7 @@ import { DataSource } from 'typeorm'
 import { UserEntity } from '../user/user-entity'
 import { ProjectEntity } from '../project/project.entity'
 import { CollectionEntity } from '../collections/collection.entity'
-import { FlowEntity } from '../flows/flow.entity'
+import { FlowEntity } from '../flows/flow/flow.entity'
 import { FlowVersionEntity } from '../flows/flow-version/flow-version-entity'
 import { FileEntity } from '../file/file.entity'
 import { StoreEntryEntity } from '../store-entry/store-entry-entity'
