+++ packages/backend/src/app/flows/flow/flow.repo.ts
@@ -1,4 +1,4 @@
-import { databaseConnection } from '../database/database-connection'
+import { databaseConnection } from '../../database/database-connection'
 import { FlowEntity } from './flow.entity'
 
 export const flowRepo = databaseConnection.getRepository(FlowEntity)
